Add EndOfFile rule
